commit before merge

+++ CoreService/src/main/java/unicauca/coreservice/domain/useCases/RubricService.java
@@ -2,8 +2,10 @@ package unicauca.coreservice.domain.useCases;
 
 import lombok.AllArgsConstructor;
 import unicauca.coreservice.application.in.RubricInt;
+import unicauca.coreservice.application.out.IAuthorizationService;
 import unicauca.coreservice.application.out.RubricRepositoryOutInt;
 import unicauca.coreservice.domain.exception.DuplicateInformation;
+import unicauca.coreservice.domain.exception.Unauthorized;
 import unicauca.coreservice.domain.model.OptionalWrapper;
 import unicauca.coreservice.domain.model.Rubric;
 import unicauca.coreservice.domain.model.SubjectOutcome;
@@ -18,10 +20,13 @@ public class RubricService implements RubricInt{
 
     private final RubricRepositoryOutInt rubricRepository;
     private final SubjectOutcomeRepository subjectOutcomeRepository;
+    private final IAuthorizationService authorizationService;
 
 
     @Override
-    public Rubric add(Rubric newRubric, Integer subjectOutcomeId) throws Exception {
+    public Rubric add(Rubric newRubric, Integer subjectOutcomeId, String uid) throws Exception {
+        if(!authorizationService.canAccessSubjectOutcome(uid, subjectOutcomeId))
+            throw new Unauthorized("You have no permission to add a rubric to this subject outcome");
         OptionalWrapper<Rubric> optionalRubric = rubricRepository.getBySubjectOutcomeId(subjectOutcomeId);
         if (optionalRubric.getValue().isPresent()) {    
             throw new DuplicateInformation(
@@ -45,31 +50,40 @@ public class RubricService implements RubricInt{
     }
 
     @Override
-    public List<Rubric> listAllBySubjectId(Integer subjectId) {
+    public List<Rubric> listAllBySubjectId(Integer subjectId, String uid) throws Exception {
+        if(!authorizationService.canAccessSubject(uid, subjectId))
+            throw new Unauthorized("You have no permission to access this subject");
         return rubricRepository.listAllBySubjectId(subjectId);
     }
 
     @Override
-    public Rubric getById(Integer id) throws Exception {
+    public Rubric getById(Integer id, String uid) throws Exception {
+        if(authorizationService.canAccessRubric(uid, id))
+            throw new Unauthorized("You have no permission to access this rubric");
         OptionalWrapper<Rubric> response = rubricRepository.getById(id);
-
         return response.getValue().orElseThrow(response::getException);
     }
 
     @Override
-    public Rubric getBySubjectOutcomeId(Integer subjectOutcomeId) throws Exception {
+    public Rubric getBySubjectOutcomeId(Integer subjectOutcomeId, String uid) throws Exception {
+        if(!authorizationService.canAccessSubjectOutcome(uid, subjectOutcomeId))
+            throw new Unauthorized("You have no permission to access this subject outcome");
         OptionalWrapper<Rubric> response = rubricRepository.getBySubjectOutcomeId(subjectOutcomeId);
         return response.getValue().orElseThrow(response::getException);
     }
 
     @Override
-    public Rubric update(Integer id, Rubric newRubric) throws Exception {
+    public Rubric update(Integer id, Rubric newRubric, String uid) throws Exception {
+        if(authorizationService.canAccessRubric(uid, id))
+            throw new Unauthorized("You have no permission to update this rubric");
         OptionalWrapper<Rubric> response = rubricRepository.update(id, newRubric);
         return response.getValue().orElseThrow(response::getException);
     }
 
     @Override
-    public Rubric remove(Integer id) throws Exception {
+    public Rubric remove(Integer id, String uid) throws Exception {
+        if(authorizationService.canAccessRubric(uid, id))
+            throw new Unauthorized("You have no permission to remove this rubric");
         OptionalWrapper<Rubric> response = rubricRepository.remove(id);
         return response.getValue().orElseThrow(response::getException);
     }
